+++ src/modulo-04/filmator-web/src/main/java/filmator/model/Filme.java
@@ -8,7 +8,7 @@ public class Filme {
 	private int anoLancamento;
 	private String sinopse;
 	private String imagem;
-	private Integer nota;
+	private Double nota;
 	
 	public Filme() {
 	}
@@ -69,11 +69,11 @@ public class Filme {
 		this.imagem = imagem;
 	}
 	
-	public Integer getNota(){
+	public Double getNota(){
 		return this.nota;
 	}
 	
-	public void setNota(Integer nota) {
+	public void setNota(Double nota) {
 		this.nota = nota;
 	}
 }
